Refactor dos Respositories
