--- src/main/kotlin/br/com/fiap/lanchonete/adapters/output/persistence/ItemPedidoRepository.kt
@@ -1,9 +1,0 @@
-package br.com.fiap.lanchonete.adapters.output.persistence
-
-import br.com.fiap.lanchonete.core.domain.entities.ItemPedido
-import org.springframework.data.jpa.repository.JpaRepository
-import org.springframework.stereotype.Repository
-import java.util.*
-
-@Repository
-interface ItemPedidoRepository : JpaRepository<ItemPedido, UUID>
