+++ src/main/kotlin/br/com/fiap/lanchonete/adapters/output/persistence/PedidoRepositoryImpl.kt
@@ -1,5 +1,6 @@
 package br.com.fiap.lanchonete.adapters.output.persistence
 
+import br.com.fiap.lanchonete.core.application.ports.output.repository.PedidoRepository
 import br.com.fiap.lanchonete.core.domain.entities.Pedido
 import br.com.fiap.lanchonete.core.domain.entities.StatusPedido
 import org.springframework.data.jpa.repository.JpaRepository
@@ -8,7 +9,7 @@ import org.springframework.stereotype.Repository
 import java.util.*
 
 @Repository
-interface PedidoRepository : JpaRepository<Pedido, UUID> {
+interface PedidoRepositoryImpl : PedidoRepository, JpaRepository<Pedido, UUID> {
     @Query("select p from Pedido p where p.status = :statusPedido")
-    fun findByStatus(statusPedido: StatusPedido): List<Pedido>
+    override fun findByStatus(statusPedido: StatusPedido): List<Pedido>
 }
