+++ src/main/kotlin/br/com/fiap/lanchonete/core/application/ports/output/repository/ClienteRepository.kt
@@ -6,4 +6,5 @@ import java.util.*
 interface ClienteRepository {
     fun save(cliente: Cliente): Cliente
     fun findByCPF(cpf: String): Optional<Cliente>
+    fun findById(id: UUID): Optional<Cliente>
 }
