--- src/main/kotlin/br/com/fiap/lanchonete/core/application/ports/output/repository/ItemPedidoRepository.kt
@@ -1,4 +1,0 @@
-package br.com.fiap.lanchonete.core.application.ports.output.repository
-
-interface ItemPedidoRepository {
-}
