+++ src/main/kotlin/br/com/fiap/lanchonete/core/application/ports/output/repository/PagamentoRepository.kt
@@ -1,4 +1,7 @@
 package br.com.fiap.lanchonete.core.application.ports.output.repository
 
+import br.com.fiap.lanchonete.core.domain.entities.Pagamento
+
 interface PagamentoRepository {
+    fun save(pagamento: Pagamento)
 }
